Fix weird nox failures with py3.9.
Any nox session that involved installing . was failing
citing strange, seemingly unrelated error messages.

Initially thought it was an openblas/numpy related thing
as this has happened to me on other projects.

However, this is fixed with an explicit call to upgrade pip,
setuptools and wheel prior to any nox sessions that installs .