User can now specify header rather than skip_rows.
Made change to API so that user can now specify header
rather than the cumbersome skip_rows.

pymechtest will now autodetect data below this header and remove it
if it is non-numeric, removing the need for skip rows.

Also because this changed the number of rows, it broke some
tests which have now been fixed and assert almost equal precision
has been changed to 2 decimal places which is more than fine
for things like strength and modulus.