Implemented autodetection in a more robust way.
Previously, autodetection was implemented in a bit
of a hacky way, with calls to 2 private methods in
every public method that touches the data.

While this worked, it wasn't exactly an elegant solution and
would likely have resulted in bugs as the project grows.

Now the stress and strain columns are implemented as
instance properties and every method that calls them
simply calls up `self.stress_col` etc.

This is implemented in the `_get_stress_strain_cols` private
method.

I've also improved the testing around this column autodetection
as well as modifying the docs to suit.